python/fundamentals/pyproject.toml
[project]
name = "fundamentals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "ipykernel>=7.0.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "orjson>=3.11.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pydantic>=2.12.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "supabase>=2.24.0",
]
